requirements.txt
Flask==2.2.5
pycryptodome==3.18.0
Werkzeug==2.2.3
